+++ apps/meteor/app/channel-settings/server/index.ts
@@ -1,5 +1,0 @@
-import './methods/saveRoomSettings';
-
-export { saveRoomTopic } from './functions/saveRoomTopic';
-export { saveRoomCategory } from './functions/saveRoomCategory';
-export { saveRoomName } from './functions/saveRoomName';