+++ apps/meteor/client/sidebar/hooks/useRoomList.ts
@@ -1,6 +1,5 @@
 import type { ILivechatInquiryRecord } from '@rocket.chat/core-typings';
 import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
-import { useFeaturePreview } from '@rocket.chat/ui-client';
 import type { SubscriptionWithRoom, TranslationKey } from '@rocket.chat/ui-contexts';
 import { useUserPreference, useUserSubscriptions, useSetting } from '@rocket.chat/ui-contexts';
 import { useVideoConfIncomingCalls } from '@rocket.chat/ui-video-conf';
@@ -21,7 +20,6 @@ const order = [
 	'Open_Livechats',
 	'On_Hold_Chats',
 	'Unread',
-	'Drafts',
 	'Favorites',
 	'Teams',
 	'Discussions',
@@ -43,7 +41,6 @@ export const useRoomList = ({ collapsedGroups }: { collapsedGroups?: string[] })
 	const showOmnichannel = useOmnichannelEnabled();
 	const sidebarGroupByType = useUserPreference('sidebarGroupByType');
 	const favoritesEnabled = useUserPreference('sidebarShowFavorites');
-	const sidebarDrafts = useFeaturePreview('sidebarDrafts');
 	const sidebarOrder = useUserPreference<typeof order>('sidebarSectionsOrder') ?? order;
 	const sidebarCustomCategoryOrder = useUserPreference<string[]>('sidebarCustomCategoryOrder') ?? [];
 	const isDiscussionEnabled = useSetting('Discussion_enabled');
@@ -63,7 +60,6 @@ export const useRoomList = ({ collapsedGroups }: { collapsedGroups?: string[] })
 		useMemo(() => {
 			const isCollapsed = (groupTitle: string) => collapsedGroups?.includes(groupTitle);
 
-			const drafts = new Set();
 			const incomingCall = new Set();
 			const favorite = new Set();
 			const omnichannel = new Set();
@@ -82,10 +78,6 @@ export const useRoomList = ({ collapsedGroups }: { collapsedGroups?: string[] })
 
 				if (sidebarShowUnread && (room.alert || room.unread || room.tunread?.length) && !room.hideUnreadStatus) {
 					return unread.add(room);
-				}
-
-				if (sidebarDrafts && room.draft) {
-					return drafts.add(room);
 				}
 
 				if (favoritesEnabled && room.f) {
@@ -132,8 +124,6 @@ export const useRoomList = ({ collapsedGroups }: { collapsedGroups?: string[] })
 			showOmnichannel && onHold.size && groups.set('On_Hold_Chats', onHold);
 
 			sidebarShowUnread && unread.size && groups.set('Unread', unread);
-
-			sidebarDrafts && drafts.size && groups.set('Drafts', drafts);
 
 			favoritesEnabled && favorite.size && groups.set('Favorites', favorite);
 
@@ -211,7 +201,6 @@ export const useRoomList = ({ collapsedGroups }: { collapsedGroups?: string[] })
 			rooms,
 			showOmnichannel,
 			inquiries.enabled,
-			sidebarDrafts,
 			queue,
 			sidebarShowUnread,
 			favoritesEnabled,