+++ packages/core-typings/src/IRoom.ts
@@ -384,6 +384,7 @@ export type RoomAdminFieldsType =
 	| 'teamMain'
 	| 'announcement'
 	| 'description'
+	| 'customFields'
 	| 'broadcast'
 	| 'uids'
 	| 'avatarETag'