+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/ui-video-conf
 
+## 33.0.0-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@33.0.0-rc.0
+  - @rocket.chat/ui-avatar@29.0.0-rc.0
+  </details>
+
+## 32.0.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-avatar@28.0.0
+  - @rocket.chat/ui-contexts@32.0.0
+  </details>
+
+## 32.0.0-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@32.0.0-rc.0
+  - @rocket.chat/ui-avatar@28.0.0-rc.0
+  </details>
+
 ## 31.0.0
 
 ### Patch Changes